Safari-611.1.10/LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-zh-Hant-HK-expected.txt

PASS navigator.language is "zh-hk"
PASS successfullyParsed is true

TEST COMPLETE
